% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenotype.R
\name{plot_phenotype}
\alias{plot_phenotype}
\title{Plot plant phenotypes and perform statistical tests}
\usage{
plot_phenotype(
  data,
  phenotype,
  plot_type = "boxplot",
  do_stats = TRUE,
  line_col = "genotype",
  treatment_col = "treatment"
)
}
\arguments{
\item{data}{A data frame resulting from \code{prepare_data}.}

\item{phenotype}{Character. The name of the column containing the phenotype to plot.}

\item{plot_type}{Character. Type of plot to create. Default is "boxplot". Other options could be "violin", etc.}

\item{do_stats}{Logical. If TRUE, perform ANOVA and post-hoc tests comparing treatments.}

\item{line_col}{Character. The name of the column that indicates the line/genotype.}

\item{treatment_col}{Character. The name of the column that indicates the treatment groups.}
}
\value{
A ggplot object.
}
\description{
This function plots the specified phenotype from the prepared data.
It can create different plot types (e.g., "boxplot") and optionally perform ANOVA and a post-hoc test.
The statistical results are displayed on the plot.
}
