% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare plant growth data}
\usage{
prepare_data(
  filepath,
  new_colnames = c("genotype", "treatment", "condition", "replicate")
)
}
\arguments{
\item{filepath}{Character. The path to the CSV file.}

\item{new_colnames}{Character vector. Names for the new columns created from splitting the "file" column.}
}
\value{
A data frame with the original data plus the new columns.
}
\description{
This function reads a CSV file containing plant growth data and splits the "file" column into multiple columns
based on underscores. The user can specify the names of these new columns.
}
